+++ tiled/adapters/jpeg.py
@@ -1,8 +1,9 @@
 import builtins
-from typing import Any, Dict, List, Optional, Tuple, Union, cast
+from typing import Any, List, Optional, Tuple, Union
 
-import tifffile
+import numpy as np
 from numpy._typing import NDArray
+from PIL import Image
 
 from ..structures.array import ArrayStructure, BuiltinDtype
 from ..structures.core import Spec, StructureFamily
@@ -13,14 +14,14 @@ from .sequence import FileSequenceAdapter
 from .type_alliases import JSON, NDSlice
 
 
-class TiffAdapter:
+class JPEGAdapter:
     """
-    Read a TIFF file.
+    Read a JPEG file.
 
     Examples
     --------
 
-    >>> TiffAdapter("path/to/file.tiff")
+    >>> JPEGAdapter("path/to/file.jpeg")
     """
 
     structure_family = StructureFamily.array
@@ -47,24 +48,17 @@ class TiffAdapter:
         if not isinstance(data_uri, str):
             raise Exception
         filepath = path_from_uri(data_uri)
-        cache_key = (tifffile.TiffFile, filepath)
-        self._file = with_resource_cache(cache_key, tifffile.TiffFile, filepath)
+        cache_key = (Image.open, filepath)
+        self._file = with_resource_cache(cache_key, Image.open, filepath)
         self.specs = specs or []
         self._provided_metadata = metadata or {}
         self.access_policy = access_policy
         if structure is None:
-            if self._file.is_shaped:
-                from_file: Tuple[Dict[str, Any], ...] = cast(
-                    Tuple[Dict[str, Any], ...], self._file.shaped_metadata
-                )
-                shape = tuple(from_file[0]["shape"])
-            else:
-                arr = self._file.asarray()
-                shape = arr.shape
+            arr = np.asarray(self._file)
             structure = ArrayStructure(
-                shape=shape,
-                chunks=tuple((dim,) for dim in shape),
-                data_type=BuiltinDtype.from_numpy_dtype(self._file.series[0].dtype),
+                shape=arr.shape,
+                chunks=tuple((dim,) for dim in arr.shape),
+                data_type=BuiltinDtype.from_numpy_dtype(arr.dtype),
             )
         self._structure = structure
 
@@ -75,11 +69,7 @@ class TiffAdapter:
         -------
 
         """
-        # This contains some enums, but Python's built-in JSON serializer
-        # handles them fine (converting  to str or int as appropriate).
-        d = {tag.name: tag.value for tag in self._file.pages[0].tags.values()}
-        d.update(self._provided_metadata)
-        return d
+        return self._provided_metadata.copy()
 
     def read(self, slice: Optional[NDSlice] = None) -> NDArray[Any]:
         """
@@ -92,17 +82,13 @@ class TiffAdapter:
         -------
 
         """
-        # TODO Is there support for reading less than the whole array
-        # if we only want a slice? I do not think that is possible with a
-        # single-page TIFF but I'm not sure. Certainly it *is* possible for
-        # multi-page TIFFs.
-        arr = self._file.asarray()
+        arr = np.asarray(self._file)
         if slice is not None:
             arr = arr[slice]
         return arr
 
     def read_block(
-        self, block: Tuple[int, ...], slice: Optional[slice] = None
+        self, block: Tuple[int, ...], slice: Optional[builtins.slice] = None
     ) -> NDArray[Any]:
         """
 
@@ -115,12 +101,10 @@ class TiffAdapter:
         -------
 
         """
-        # For simplicity, this adapter always treat a single TIFF file as one
-        # chunk. This could be relaxed in the future.
         if sum(block) != 0:
             raise IndexError(block)
 
-        arr = self._file.asarray()
+        arr = np.asarray(self._file)
         if slice is not None:
             arr = arr[slice]
         return arr
@@ -135,8 +119,15 @@ class TiffAdapter:
         return self._structure
 
 
-class TiffSequenceAdapter(FileSequenceAdapter):
+class JPEGSequenceAdapter(FileSequenceAdapter):
     def _load_from_files(
         self, slice: Union[builtins.slice, int] = slice(None)
     ) -> NDArray[Any]:
-        return tifffile.TiffSequence(self.filepaths[slice]).asarray()
+        from PIL import Image
+
+        if isinstance(slice, int):
+            return np.asarray(Image.open(self.filepaths[slice]))[None, ...]
+        else:
+            return np.asarray(
+                [np.asarray(Image.open(file)) for file in self.filepaths[slice]]
+            )