+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Write the date in place of the "Unreleased" in the case a new version is release
 - Automatic reshaping of tiff data by the adapter to account for
   extra/missing singleton dimensions
 - Refactor CSVAdapter to allow pd.read_csv kwargs
+- Add adapters for reading back assets with the image/jpeg and
+  multipart/related;type=image/jpeg mimetypes.
 
 ## v0.1.0b10 (2024-10-11)